Default VOUCH_PORT
I found trying to run v0.17.3 that I had to set this
env var as it was otherwise defaulted to a hostname
which vouch failed to parse as an int on start up.

Also, had to remove `webapp: false` from the config
as this conflicted with other settings and vouch
failed to start.