it is safer if the jwt.secret is not set by default
This change probably warrents attention from the user, as it may
invalidate the sessions of logged in users when this change is
uptaken.

However I believe it is an acceptable interference. As things are
now, if a deployment of vouch does not set the jwt.secret then the
cookie, copied to another domain would be accepted.

With this change, the default behaviour causes vouch to generate a
secret.